// ui/src/app/model/api/simplePayrollRuleItem.ts
/**
 * Balancing Act API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: balancing-act@moroski.info
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { ObjectId } from './objectId';
import { PayrollRuleItem } from './payrollRuleItem';


export interface SimplePayrollRuleItem extends PayrollRuleItem {
    amount?: number;
}